orches/requirements.txt
sqlalchemy>=2.0
python-dotenv>=1.0
fastapi>=0.111
uvicorn[standard]>=0.29
anthropic>=0.28
openai>=1.30
httpx>=0.27
pydantic>=2.7
websockets>=12.0
ddgs>=0.1
apscheduler>=3.10
croniter>=2.0
mcp-server-fetch>=0.1
aiofiles>=23.0
